requirements.txt
altgraph==0.17
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cryptography==2.8
decorator==4.4.1
distlib==0.3.0
Django==3.0.3
django-grappelli==2.14.1
django-modeltranslation==0.14.4
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.9
pefile==2019.4.18
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
PyInstaller==3.6
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.4.2
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.8
virtualenv==20.0.10
virtualenvwrapper-win==1.2.6
